requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
googleapis_common_protos
google-cloud-ndb